pyproject.toml
[tool.poetry]
name = "electronic_network_project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
redis = "^5.0.14"
django-filter = "^25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"